# openapi.yaml
openapi: 3.1.0
info:
  title: National Transit Connect API
  version: 0.1.0
  summary: REST + SSE backend for bus routes, trips, live locations, and alerts.
  description: |-
    This document describes the HTTP and Server-Sent Events surfaces exposed by the
    National Transit Connect API. It covers authentication, pagination, conditional
    caching, rate limiting, and resource models for routes, buses, trips, alerts,
    and live bus location updates.
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.example.com
    description: Placeholder production hostname
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      description: 1-based page number. Defaults to 1 when omitted.
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Page size. Defaults to 20. Maximum varies per resource (commonly 100).
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 20
    SortParam:
      name: sort
      in: query
      description: "Sort expression supporting dotted fields and '-' prefix for descending order (e.g. `-updatedAt`)."
      schema:
        type: string
      example: "-updatedAt"
    FilterParam:
      name: filter
      in: query
      description: Deep object filter (e.g. `filter[name]=North`, `filter[status]=active`).
      style: deepObject
      explode: true
      schema:
        type: object
        additionalProperties:
          type: string
    IfNoneMatchHeader:
      name: If-None-Match
      in: header
      description: Provide the ETag from a previous response to leverage conditional caching (304 Not Modified when unchanged).
      schema:
        type: string
    IfModifiedSinceHeader:
      name: If-Modified-Since
      in: header
      description: Provide the Last-Modified value from a previous response to leverage conditional caching.
      schema:
        type: string
        format: date-time
  headers:
    ETag:
      description: Strong ETag representing the response entity. Reuse with `If-None-Match` on subsequent requests.
      schema:
        type: string
    LastModified:
      description: Timestamp identifying the last mutation of the resource collection or entity.
      schema:
        type: string
        format: date-time
    RateLimitLimit:
      description: Maximum requests permitted in the current window.
      schema:
        type: integer
    RateLimitRemaining:
      description: Requests remaining in the current window.
      schema:
        type: integer
    RateLimitReset:
      description: UNIX epoch seconds when the current window resets.
      schema:
        type: integer
        format: int64
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: integer
              format: int32
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    Route:
      type: object
      properties:
        id:
          type: string
          description: MongoDB ObjectId of the route
          example: 651f8dd7f82f10065b4a8c90
        code:
          type: string
          example: RX01
        name:
          type: string
        provinceFrom:
          type: string
        provinceTo:
          type: string
        distanceKm:
          type: number
          format: float
          example: 180
        ownerId:
          type: string
          description: Owning operator's user id (for RBAC checks)
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - code
        - name
        - provinceFrom
        - provinceTo
        - distanceKm
    PaginatedRouteList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Route'
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
      required:
        - data
        - page
        - limit
        - total
...
    Trip:
      type: object
      properties:
        id:
          type: string
        routeId:
          type: string
        busId:
          type: string
        serviceDate:
          type: string
          format: date
        schedDepart:
          type: string
          format: date-time
        schedArrive:
          type: string
          format: date-time
        status:
          type: string
          description: Trip lifecycle status (scheduled, enroute, completed, cancelled)
        ownerId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - routeId
        - busId
        - serviceDate
        - schedDepart
        - schedArrive
    PaginatedTripList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Trip'
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
      required:
        - data
        - page
        - limit
        - total
    Alert:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        body:
          type: string
        type:
          type: string
          description: Audience or channel type (e.g. system, route-specific)
        severity:
          type: string
          enum: [info, warning, critical]
        routes:
          type: array
          items:
            type: string
          description: Route ids impacted by the alert
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - body
        - severity
    PaginatedAlertList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Alert'
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
      required:
        - data
        - page
        - limit
        - total
...
paths:
...
  /api/v1/routes:
    get:
      tags: [Routes]
      summary: List routes with pagination and filtering
      operationId: listRoutes
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/FilterParam'
        - $ref: '#/components/parameters/IfNoneMatchHeader'
        - $ref: '#/components/parameters/IfModifiedSinceHeader'
      security: []
      responses:
        '200':
          description: Routes returned successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Last-Modified:
              $ref: '#/components/headers/LastModified'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedRouteList'
        '304':
          description: Not modified since last request
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [Routes]
      summary: Create a new route
      operationId: createRoute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, name, provinceFrom, provinceTo, distanceKm]
              properties:
                code:
                  type: string
                name:
                  type: string
                provinceFrom:
                  type: string
                provinceTo:
                  type: string
                distanceKm:
                  type: number
                  format: float
                ownerId:
                  type: string
      responses:
        '201':
          description: Route created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Duplicate route code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/routes/by-code/{code}:
    get:
      tags: [Routes]
      summary: Retrieve a route by its short code
      operationId: getRouteByCode
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/IfNoneMatchHeader'
        - $ref: '#/components/parameters/IfModifiedSinceHeader'
      security: []
      responses:
        '200':
          description: Route found
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Last-Modified:
              $ref: '#/components/headers/LastModified'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '404':
          description: Route not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/routes/{id}:
    get:
      tags: [Routes]
      summary: Retrieve a route by id
      operationId: getRoute
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/IfNoneMatchHeader'
        - $ref: '#/components/parameters/IfModifiedSinceHeader'
      security: []
      responses:
        '200':
          description: Route found
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Last-Modified:
              $ref: '#/components/headers/LastModified'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '404':
          description: Route not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags: [Routes]
      summary: Update editable fields on a route
      operationId: updateRoute
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                provinceFrom:
                  type: string
                provinceTo:
                  type: string
                distanceKm:
                  type: number
                  format: float
      responses:
        '200':
          description: Route updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Route not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Routes]
      summary: Delete a route
      operationId: deleteRoute
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Route removed
        '404':
          description: Route not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/trips:
    get:
      tags: [Trips]
      summary: List trips with pagination and filtering
      operationId: listTrips
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/FilterParam'
        - $ref: '#/components/parameters/IfNoneMatchHeader'
        - $ref: '#/components/parameters/IfModifiedSinceHeader'
      security: []
      responses:
        '200':
          description: Trips returned successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Last-Modified:
              $ref: '#/components/headers/LastModified'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTripList'
        '304':
          description: Not modified since last request
    post:
      tags: [Trips]
      summary: Create a new trip assignment
      operationId: createTrip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [routeId, busId, serviceDate, schedDepart, schedArrive]
              properties:
                routeId:
                  type: string
                busId:
                  type: string
                serviceDate:
                  type: string
                  format: date
                schedDepart:
                  type: string
                  format: date-time
                schedArrive:
                  type: string
                  format: date-time
                ownerId:
                  type: string
      responses:
        '201':
          description: Trip created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trip'
        '400':
          description: Validation error or scheduling conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Route or bus not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/trips/{id}:
    get:
      tags: [Trips]
      summary: Retrieve a trip by id
      operationId: getTrip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/IfNoneMatchHeader'
        - $ref: '#/components/parameters/IfModifiedSinceHeader'
      security: []
      responses:
        '200':
          description: Trip found
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Last-Modified:
              $ref: '#/components/headers/LastModified'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trip'
        '404':
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags: [Trips]
      summary: Update a trip schedule or metadata
      operationId: updateTrip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                serviceDate:
                  type: string
                  format: date
                schedDepart:
                  type: string
                  format: date-time
                schedArrive:
                  type: string
                  format: date-time
                status:
                  type: string
                ownerId:
                  type: string
      responses:
        '200':
          description: Trip updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trip'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Trips]
      summary: Delete a trip
      operationId: deleteTrip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Trip removed
        '404':
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/alerts:
    get:
      tags: [Alerts]
      summary: List service alerts
      operationId: listAlerts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/FilterParam'
        - $ref: '#/components/parameters/IfNoneMatchHeader'
        - $ref: '#/components/parameters/IfModifiedSinceHeader'
      security: []
      responses:
        '200':
          description: Alerts returned successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Last-Modified:
              $ref: '#/components/headers/LastModified'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAlertList'
        '304':
          description: Not modified since last request
    post:
      tags: [Alerts]
      summary: Publish a new alert
      operationId: createAlert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, body, severity]
              properties:
                title:
                  type: string
                body:
                  type: string
                severity:
                  type: string
                  enum: [info, warning, critical]
                type:
                  type: string
                routes:
                  type: array
                  items:
                    type: string
                startAt:
                  type: string
                  format: date-time
                endAt:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Alert created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/alerts/{id}:
    get:
      tags: [Alerts]
      summary: Retrieve an alert by id
      operationId: getAlert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/IfNoneMatchHeader'
        - $ref: '#/components/parameters/IfModifiedSinceHeader'
      security: []
      responses:
        '200':
          description: Alert found
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Last-Modified:
              $ref: '#/components/headers/LastModified'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
        '404':
          description: Alert not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags: [Alerts]
      summary: Update an alert
      operationId: updateAlert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                body:
                  type: string
                severity:
                  type: string
                  enum: [info, warning, critical]
                type:
                  type: string
                routes:
                  type: array
                  items:
                    type: string
                startAt:
                  type: string
                  format: date-time
                endAt:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Alert updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
        '404':
          description: Alert not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Alerts]
      summary: Delete an alert
      operationId: deleteAlert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Alert deleted
        '404':
          description: Alert not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...